refactor(test_helpers): moved scenarios functions to test_helpers
now that constants are not defined in common, this helper can easily be
used in common too.
