+++ core/common/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod errors;
 pub mod fs;
-pub mod test_helpers;
 
 use constants::{PENDING_SIGNERS_DIR, SIGNERS_DIR, SIGNERS_FILE, SIGNERS_HISTORY_FILE};
 use serde::{Deserialize, Serialize};
@@ -11,6 +10,8 @@ use std::fs::File;
 use std::marker::PhantomData;
 use std::path::Path;
 use std::str::FromStr;
+
+use crate::fs::names::find_global_signers_for;
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(try_from = "String", into = "String")]
@@ -271,8 +272,6 @@ impl SignedFileLoader {
 
 #[cfg(test)]
 mod asfaload_common_tests {
-    use crate::test_helpers::scenarios::setup_asfald_project_registered;
-
     use super::*;
     use anyhow::Result;
     use sha2::{Digest, Sha512};
@@ -280,6 +279,7 @@ mod asfaload_common_tests {
     use std::io::Write;
     use std::path::Path;
     use tempfile::{NamedTempFile, TempDir};
+    use test_helpers::scenarios::setup_asfald_project_registered;
     //
     // Helper to convert byte array to hex string
     fn to_hex(bytes: &[u8]) -> String {
@@ -356,7 +356,9 @@ mod asfaload_common_tests {
 
     #[test]
     fn test_determine_file_type_github_hierarchy() -> Result<()> {
-        let pending_index_file = setup_asfald_project_registered()?;
+        let temp_dir = TempDir::new()?;
+        let pending_index_file =
+            setup_asfald_project_registered(temp_dir.path().to_path_buf(), "{}")?;
         // This should be InitialSigners because there's NO asfaload.signers dir
         // in asfald/ or any parent directory
         assert_eq!(
