+++ core/test_helpers/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2024"
 [dependencies]
 signatures = { path = "../signatures" }
 minisign = "0.8.0"
+constants = { path = "../constants/" }
+anyhow = "1.0.100"
 
 [lints]
 workspace = true
